requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
langchain==0.3.21
langchain-openai==0.3.9
